requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-genai
google-genai
python-multipart
